backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
arabic-reshaper==3.0.0
asgiref==3.9.1
asn1crypto==1.5.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
Brotli==1.1.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
coverage==7.10.6
cryptography==45.0.5
cssselect2==0.8.0
cycler==0.12.1
defusedxml==0.7.1
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.27.2
et_xmlfile==2.0.0
fastapi==0.116.1
Flask==3.1.1
fonttools==4.59.0
fpdf==1.7.2
fpdf2==2.8.3
greenlet==3.2.3
h11==0.16.0
html5lib==1.1
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.8
lxml==6.0.0
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
narwhals==1.48.1
numpy==2.3.2
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pandas==2.3.1
pdf2image==1.17.0
pillow==11.3.0
playwright==1.54.0
plotly==6.2.0
pluggy==1.6.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydyf==0.11.0
pyee==13.0.0
Pygments==2.19.2
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
PyJWT==2.10.1
pyparsing==3.2.3
pypdf==5.8.0
pyphen==0.17.2
pytest==8.4.2
pytest-django==4.11.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
reportlab==4.4.3
requests==2.32.4
rich==14.1.0
rpds-py==0.27.1
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
starlette==0.47.2
svglib==1.5.1
tabulate==0.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
uritools==5.0.0
urllib3==2.5.0
uvicorn==0.35.0
weasyprint==66.0
webencodings==0.5.1
Werkzeug==3.1.3
xhtml2pdf==0.2.17
zopfli==0.2.3.post1
